feat: crypto tutorial (#273)
* A cool tutorial

This tutorial contains an example with multiple indexes and subsequent tampering to the data in order to show resiliency and a real life use case of TC applied on cryptocurrency prices up to 2021.

* Change merging method 

Fix to the issue raised on empty dataframe, resulting from an inner merge where the existing dataframe that accumulated results and the newer one had diferent indices. A subsequent issue must be raised to either: 
1 report a single model failure (on index matching) 
2 fix the moirai discrepancy (only model that showed this issue)

* run pre-comit hooks

* Changing filename to requested formats and folder. This also contains the title suggested

* Clearance of verbose cells

* Staging changes to mkdocs and the naming of the tutorial file

---------

Co-authored-by: azul <fede.garza.ramirez@gmail.com>
Co-authored-by: elmartinj <elmartinj>